Hopefully this fixes the crash bug when autonomous activators use silkworms to crash the server...

+++ ExNihilo/code/exnihilo/items/ItemSilkworm.java
@@ -19,51 +19,58 @@ public class ItemSilkworm extends Item{
 		super(par1);
 		setCreativeTab(CreativeTabs.tabMisc);
 	}
-	
+
 	/**
-    * Callback for item usage. If the item does something special on right clicking, he will have one of those. Return
-    * True if something happen and false if it don't. This is for ITEMS, not BLOCKS
-    */
-   public boolean onItemUse(ItemStack item, EntityPlayer player, World world, int x, int y, int z, int par7, float par8, float par9, float par10)
-   {
-	   Block block = Block.blocksList[world.getBlockId(x, y, z)];
-	   if (block.isLeaves(null, 0, 0, 0) && !Forestry.addsThisLeaf(block) && block.blockID != Blocks.LeavesInfested.blockID)
-	   {
-		   int oldID = world.getBlockId(x, y, z);
-		   int oldMeta = world.getBlockMetadata(x, y, z);
-		   
-		   world.setBlock(x, y, z, Blocks.LeavesInfested.blockID, oldMeta, 3);
-		   TileEntityLeavesInfested te = (TileEntityLeavesInfested)world.getBlockTileEntity(x, y, z); 
-		   te.setMimicBlock(oldID, oldMeta);
-		   
-		   item.stackSize -= 1;
-		   
-		   if (item.stackSize <= 0)
-		   {
-			  player.destroyCurrentEquippedItem(); 
-		   }
-		   
-		   return true;
-	   }
-	   else
-	   {
-		   return false;
-	   }
-		   
-   }
+	 * Callback for item usage. If the item does something special on right clicking, he will have one of those. Return
+	 * True if something happen and false if it don't. This is for ITEMS, not BLOCKS
+	 */
+	public boolean onItemUse(ItemStack item, EntityPlayer player, World world, int x, int y, int z, int par7, float par8, float par9, float par10)
+	{
+		if (!world.isAirBlock(x, y, z))
+		{
+			Block block = Block.blocksList[world.getBlockId(x, y, z)];
+			if (block.isLeaves(null, 0, 0, 0) && !Forestry.addsThisLeaf(block) && block.blockID != Blocks.LeavesInfested.blockID)
+			{
+				int oldID = world.getBlockId(x, y, z);
+				int oldMeta = world.getBlockMetadata(x, y, z);
+
+				world.setBlock(x, y, z, Blocks.LeavesInfested.blockID, oldMeta, 3);
+				TileEntityLeavesInfested te = (TileEntityLeavesInfested)world.getBlockTileEntity(x, y, z); 
+				te.setMimicBlock(oldID, oldMeta);
+
+				item.stackSize -= 1;
+
+				if (item.stackSize <= 0)
+				{
+					player.destroyCurrentEquippedItem(); 
+				}
+
+				return true;
+			}
+			else
+			{
+				return false;
+			}
+		}
+		else
+		{
+			return false;
+		}
+
+	}
 
 	@Override
 	public String getUnlocalizedName()
 	{
 		return ItemData.SILKWORM_UNLOCALIZED_NAME;
 	}
-	
+
 	@Override
 	public String getUnlocalizedName(ItemStack item)
 	{
 		return ItemData.SILKWORM_UNLOCALIZED_NAME;
 	}
-	
+
 	@Override
 	public void registerIcons(IconRegister register)
 	{
